feat(docker): adding dev dockerfile + minor folder restructure
