+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './layout';
-export { default as Routes } from './Routes';
