+++ src/utils/test-utils.tsx
@@ -1,5 +1,3 @@
-import { initialState } from '@store';
-import { rootSaga } from '@store/sagas';
 import { render, RenderResult } from '@testing-library/react';
 import { defaultTheme } from '@theme';
 import axios from 'axios';
@@ -9,6 +7,8 @@ import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import reduxMockStore from 'redux-mock-store';
 import reduxSaga from 'redux-saga';
+import { initialState } from 'state';
+import { rootSaga } from 'state/sagas';
 import { ThemeProvider } from './styled';
 
 export const mockAxios = new MockAdapter(axios);
@@ -61,3 +61,4 @@ export const renderWithStoreAndRouter = (ui: React.ReactElement, options?: objec
 export * from '@testing-library/react';
 // override render method
 export { customRender as render, store };
+
